added motor intake subsytem, implemented launcher gate

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/fy25/subsystems/launchergate/LauncherGateBlank.java
@@ -24,7 +24,7 @@ public class LauncherGateBlank implements LauncherGate {
     }
 
     @Override
-    public int getPower() {
+    public double getPower() {
         return 0;
     }
