+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/fy25/subsystems/launchergate/LauncherGateServoImpl.java
@@ -27,7 +27,7 @@ public class LauncherGateServoImpl implements LauncherGate {
     }
 
     @Override
-    public int getPower() {
+    public double getPower() {
         return 0;
     }
